backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.11
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.2
Pillow==9.4.0
pycparser==2.21
pyhcl==0.4.4
pytz==2022.7.1
requests==2.28.2
shortuuid==1.0.11
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
validators==0.20.0
win32-setctime==1.1.0
